tested and edited code for grabber
continued the work from last meeting. Tested the code that was written and then made modifications to it until the grabber was able to go to roughly 45% open each time and hold the postition. the grabber should exit the endcoder postition once found. comments were also added to ry to help it make sense