Timing changes to autonomous
Deleted the unused code to make it easier to find and edit the necessary functions. Added to the autonomous functions so that time can be editied for the chassis and lift for each individual section of code (scorePresetCone & scoreLeftSide). Timings were edited to adjust for tests and redundnacies were removed