More autonomous + grabber edits
More edits were made in autnomous, making the timing get closer to 15 seconds total, and to make it easier to edit. Edits were also made to the driver code for the grabber in attempts to make it run smooother. This has yet to be proven, since the claw has stopped working