Mid tournament auton changes
Having fresh batteries really makes a difference. Had to change some times to adjust for the charge. Also, we got selected by another team to go into the tournament and we won the whole  tournament. This team also scored on the sationary goal, hence the big wait at the beginning of autonomous